+++ lib/models/types.dart
@@ -1,10 +1,7 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/widgets.dart' show FormFieldState, GlobalKey, FormState;
 
 /// A type alias for a form key.
 typedef FormKey = GlobalKey<FormState>;
-
-/// A type alias for a validator function.
-typedef ValidatorFn<T> = String? Function(T? value, BuildContext context);
 
 /// A type alias for a validator function.
 typedef ValidatorFn2<T> = String? Function(T? value);